# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Audiobook Librarian Lite API",
    "version": "1.0.0",
    "description": "API for the Audiobook Librarian **lite** server: a self-hosted sync backend for the Audiobook Librarian apps.\n\nLite stores listening events, playback positions, bookmarks, tags, goals and badges for its own users, and nothing else. It hosts **no book catalog** — no books, authors, series, genres, covers, downloads or imports — and it never contacts or shares data with a full Audiobook Librarian server.\n\n**Book identity.** Because there is no catalog there are no numeric book ids. Every endpoint that refers to a book identifies it by the pair `title` + `author`, sent as query parameters or body fields. Clients built against the full server may still send `bookId` on sync payloads; it is accepted and ignored.\n\nAuthentication uses opaque bearer tokens returned by the login and OTP endpoints. An account created by self-service signup starts at role `unverified` and is refused (403) by every authenticated endpoint until an administrator approves it.",
    "contact": {
      "name": "API Support",
      "email": "support@librarian.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://{host}/api/v1",
      "description": "Your self-hosted lite server",
      "variables": {
        "host": {
          "default": "localhost:8000"
        }
...
    }
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Authentication using Bearer tokens. After logging in, include the token in the Authorization header: `Authorization: Bearer {token}`"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code for identifying the type of error",
            "example": "ACCOUNT_PENDING_APPROVAL"
          },
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Unauthorized"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message",
            "example": "Invalid credentials"
          }
        }
      },
...
      "ListeningEvent": {
        "type": "object",
        "description": "A single playback event. Lite hosts no book catalog, so an event is keyed by `bookTitle` + `bookAuthor` rather than by a numeric book id. Clients built against the full server may instead send `bookId` with the same strings carried in `metadata.fallbackTitle` / `metadata.fallbackAuthor`; lite reads the identity from there and ignores the id. Responses always carry `bookTitle` / `bookAuthor`.",
        "required": [
          "id",
          "eventType",
          "timestampMs",
          "positionMs",
          "deviceId",
          "timezone",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Client-generated event id; re-sending a known id is a no-op.",
            "maxLength": 255
          },
          "bookTitle": {
            "type": "string",
            "maxLength": 255,
            "example": "Dune"
          },
          "bookAuthor": {
            "type": "string",
            "maxLength": 255,
            "example": "Frank Herbert"
          },
          "bookPath": {
            "type": "string",
            "nullable": true,
            "maxLength": 500,
            "description": "Client-local path. Stored for the client's benefit only; never used as identity."
          },
          "bookId": {
            "type": "integer",
            "nullable": true,
            "deprecated": true,
            "description": "Accepted from full-server clients and ignored. Lite identifies books by title and author."
          },
          "eventType": {
            "type": "string",
            "maxLength": 50,
            "example": "SESSION_END"
          },
          "timestampMs": {
            "type": "integer",
            "minimum": 0
          },
          "positionMs": {
            "type": "integer",
            "minimum": 0
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "deviceId": {
            "type": "string",
            "maxLength": 255
          },
          "timezone": {
            "type": "string",
            "maxLength": 50
          },
          "syncStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "minimum": 0
          },
          "syncedAt": {
            "type": "integer",
            "nullable": true
          },
          "migratedFrom": {
            "type": "string",
            "nullable": true,
            "maxLength": 50
          },
          "migrationSourceId": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/sync/events/book": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Get book events",
        "description": "Get listening events for a specific book, identified by title/author.",
        "operationId": "getBookEvents",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "author",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ListeningEvent"
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "hasMore": {
                      "type": "boolean"
                    },
                    "title": {
                      "type": "string"
                    },
                    "author": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}